extract node.h
