+++ bin/template
@@ -193,9 +193,9 @@ nodes = config.fetch("nodes").map { |node| NodeType.new(node) }.sort_by(&:name)
 tokens = config.fetch("tokens").map { |token| Token.new(token) }
 
 template("../ext/yarp/extension.c", nodes, cmark)
-template("../ext/yarp/yarp.h", nodes, cmark)
 template("../ext/yarp/yarp.c", nodes, cmark)
 template("../lib/yarp.rb", nodes, rbmark)
 
 pure_template(template: 'bin/templates/token_type.h.erb', write_to: 'ext/yarp/token_type.h', locals: { tokens: tokens })
 pure_template(template: 'bin/templates/token_type.c.erb', write_to: 'ext/yarp/token_type.c', locals: { tokens: tokens })
+pure_template(template: 'bin/templates/node.h.erb', write_to: 'ext/yarp/node.h', locals: { nodes: nodes })
